@@ -1,16 +1,15 @@
-
 package servlet;
 
 import fabrica.FabricaNegocio;
 import fabrica.IFabricaNegocio;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.util.List;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import web.blogdominio.domain.Anclada;
 import web.blogdominio.domain.Comun;
 import web.blogdominio.domain.Usuario;
@@ -19,8 +18,8 @@ import web.blogdominio.domain.Usuario;
  *
  * @author jairo-rhz
  */
+@WebServlet(name = "SvLogin", urlPatterns = {"/SvLogin"})
 public class SvLogin extends HttpServlet {
-
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
@@ -29,25 +28,32 @@ public class SvLogin extends HttpServlet {
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String avatar = request.getParameter("avatar");
+        String email = request.getParameter("email");
         String contrasenia = request.getParameter("contrasenia");
+
+        System.out.println("holaaaaa");
+        System.out.println(email);
 
         IFabricaNegocio fabricaNegocio = new FabricaNegocio();
         Usuario usuario = fabricaNegocio.createUsuarioNegocio().
-                validarInicioUsuario(avatar, contrasenia);
+                validarInicioUsuario(email, contrasenia);
         List<Comun> publicacionesComunes = fabricaNegocio.createPublicacionNegocio().consultarPublicacionesComunes();
         List<Anclada> publicacionesAncladas = fabricaNegocio.createPublicacionNegocio().consultarPublicacionesAncladas();
+        
+//        for (Anclada publicacionAnclada : publicacionesAncladas) {
+//           System.out.println(publicacionAnclada.toString());
+//       }
         if (usuario != null) {
 
             HttpSession sesion = request.getSession();
             sesion.setAttribute("usuario", usuario);
             sesion.setAttribute("publicacionesComunes", publicacionesComunes);
             sesion.setAttribute("publicacionesAncladas", publicacionesAncladas);
-            response.sendRedirect(request.getContextPath() + "/paginas/PaginaPrincipal.jsp");
+            response.sendRedirect(request.getContextPath() + "/home.jsp");
 
-        }else{
-            response.sendRedirect("errores.jsp");
-            String msg="Favor de ingresar un usuario válido";
+        } else {
+
+            response.sendRedirect(request.getContextPath() + "/errores.jsp");
         }
     }
 }
