@@ -1,17 +1,16 @@
-
 package servlet;
 
 import fabrica.FabricaNegocio;
 import fabrica.IFabricaNegocio;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.util.Calendar;
 import java.util.List;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import web.blogdominio.domain.Anclada;
 import web.blogdominio.domain.Comun;
 import web.blogdominio.domain.Credencial;
@@ -24,6 +23,7 @@ import web.blogdominio.domain.Normal;
  *
  * @author jairo-rhz
  */
+@WebServlet(name = "SvRegister", urlPatterns = {"/SvRegister"})
 public class SvRegister extends HttpServlet {
  @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
@@ -42,25 +42,30 @@ public class SvRegister extends HttpServlet {
         String municipio = request.getParameter("municipio");
         String fechaNacimiento = request.getParameter("fechaNacimiento");
         String genero = request.getParameter("genero");
-        String usuario = request.getParameter("usuario");
+        //String usuario = request.getParameter("usuario");
         String contrasenia = request.getParameter("contrasenia");
 
         Normal usuarioNormal = new Normal();
         usuarioNormal.setNombreCompleto(nombreCompleto);
-        usuarioNormal.setCredencial(new Credencial(usuario, contrasenia));
+        usuarioNormal.setCredencial(new Credencial(email, contrasenia));
         usuarioNormal.setTelefono(telefono);
         usuarioNormal.setCorreo(email);
         usuarioNormal.setMunicipio(
                 new Municipio(municipio, new Estado(estado)));
         usuarioNormal.setCiudad(ciudad);
-        if (genero.equals("Hombre")) {
+        if (genero.equals("m")) {
             usuarioNormal.setGenero(Genero.MASCULINO);
-        } else {
+        } else if (genero.equals("f")){
             usuarioNormal.setGenero(Genero.FEMENINO);
+        }
+        if (genero.equals("o")){
+            usuarioNormal.setGenero(Genero.OTRO);
         }
 
         usuarioNormal.setFechaNacimiento(Calendar.getInstance());
-
+        
+        
+        
         IFabricaNegocio fabricaNegocio = new FabricaNegocio();
 
         Normal UsuarioNormalNuevo = (Normal) fabricaNegocio.createUsuarioNegocio().registrarUsuario(usuarioNormal);
@@ -71,7 +76,7 @@ public class SvRegister extends HttpServlet {
             sesion.setAttribute("usuario", UsuarioNormalNuevo);
             sesion.setAttribute("publicacionesComunes", publicacionesComunes);
             sesion.setAttribute("publicacionesAncladas", publicacionesAncladas);
-            response.sendRedirect(request.getContextPath() + "/paginas/PaginaPrincipal.jsp");
+            response.sendRedirect(request.getContextPath() + "/home.jsp");
 
         }
 //        response.sendRedirect("/paginas/Login.jsp");
